requirements.txt
numpy
scipy
scikit-learn
matplotlib
openai>=1.3.0
python-dotenv
streamlit
plotly
pymupdf